# docs/api/openapi-v1.yaml
...
openapi: 3.1.0
info:
  title: PigOS API
  description: 'Global Swine Farm Management SaaS — Base API + AI Addon platform.


    이 파일은 라이브 라우트에서 자동 생성됨 (scripts/gen_openapi.py). 수기 편집 금지.

    사람용 연동 계약서: docs/mobile-integration-contract.md


    인증: Bearer JWT (/api/v1/auth/login). access 15분 / refresh 7일.

    멀티테넌트: 모든 농장 데이터는 farm_id로 격리.'
  version: 0.1.0
  contact:
    name: PigOS Support
    url: https://pigos.io
  license:
    name: Proprietary
paths:
...
  /api/v1/farms/{farm_id}/events/definitions:
    get:
      tags:
      - Events
      summary: List Event Definitions
      description: 'Return event types applicable to this farm''s country.

        Filters regional_applicability: ALL always included;

        comma-separated ISO codes included only if farm.country matches.

        Phase filter: MVP only.'
      operationId: list_event_definitions_api_v1_farms__farm_id__events_definitions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDefinitionResponse'
                title: Response List Event Definitions Api V1 Farms  Farm Id  Events
                  Definitions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/events/matings:
    get:
      tags:
      - Events
      summary: List Matings
      operationId: list_matings_api_v1_farms__farm_id__events_matings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: sow_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Sow Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatingResponse'
                title: Response List Matings Api V1 Farms  Farm Id  Events Matings
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Events
      summary: Record Mating
      description: 'Record a mating event. Automatically:

        - Creates or reuses a BreedingCycle

        - Sets sow.status = GESTATING'
      operationId: record_mating_api_v1_farms__farm_id__events_matings_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatingCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/events/farrowings:
    get:
      tags:
      - Events
      summary: List Farrowings
      operationId: list_farrowings_api_v1_farms__farm_id__events_farrowings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: sow_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Sow Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FarrowingResponse'
                title: Response List Farrowings Api V1 Farms  Farm Id  Events Farrowings
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Events
      summary: Record Farrowing
      description: 'Record a farrowing. Validates gestation period (100–130 days).

        Sets sow.status = LACTATING, increments parity.'
      operationId: record_farrowing_api_v1_farms__farm_id__events_farrowings_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarrowingCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarrowingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/events/weanings:
    get:
      tags:
      - Events
      summary: List Weanings
      operationId: list_weanings_api_v1_farms__farm_id__events_weanings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: sow_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Sow Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeaningResponse'
                title: Response List Weanings Api V1 Farms  Farm Id  Events Weanings
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Events
      summary: Record Weaning
      description: 'Record a weaning. Validates nursing period (10–60 days).

        Sets sow.status = WEANED, closes BreedingCycle.'
      operationId: record_weaning_api_v1_farms__farm_id__events_weanings_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeaningCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeaningResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/events/reproductive:
    post:
      tags:
      - Events
      summary: Record Reproductive Event
      description: 'Record non-productive event: return to estrus, abortion, empty,
        cull, death.

        Critical for accurate NPD calculation.'
      operationId: record_reproductive_event_api_v1_farms__farm_id__events_reproductive_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReproductiveEventCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReproductiveEventResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/events/piglet_events:
    get:
      tags:
      - Events
      summary: List Piglet Events
      operationId: list_piglet_events_api_v1_farms__farm_id__events_piglet_events_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: sow_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Sow Id
      - name: farrowing_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Farrowing Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PigletEventResponse'
                title: Response List Piglet Events Api V1 Farms  Farm Id  Events Piglet
                  Events Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Events
      summary: Record Piglet Event
      description: 'Record an in-lactation piglet event: death, foster in/out.

        farrowing_id is optional — auto-selects most recent farrowing if omitted.'
      operationId: record_piglet_event_api_v1_farms__farm_id__events_piglet_events_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PigletEventCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PigletEventResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/events/matings/{mating_id}:
    patch:
      tags:
      - Events
      summary: Update Mating
      operationId: update_mating_api_v1_farms__farm_id__events_matings__mating_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: mating_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Mating Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatingUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Events
      summary: Delete Mating
      operationId: delete_mating_api_v1_farms__farm_id__events_matings__mating_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: mating_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Mating Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/events/farrowings/{farrowing_id}:
    patch:
      tags:
      - Events
      summary: Update Farrowing
      operationId: update_farrowing_api_v1_farms__farm_id__events_farrowings__farrowing_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: farrowing_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Farrowing Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FarrowingUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FarrowingResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Events
      summary: Delete Farrowing
      operationId: delete_farrowing_api_v1_farms__farm_id__events_farrowings__farrowing_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: farrowing_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Farrowing Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/events/weanings/{weaning_id}:
    patch:
      tags:
      - Events
      summary: Update Weaning
      operationId: update_weaning_api_v1_farms__farm_id__events_weanings__weaning_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: weaning_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Weaning Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeaningUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeaningResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Events
      summary: Delete Weaning
      operationId: delete_weaning_api_v1_farms__farm_id__events_weanings__weaning_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: weaning_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Weaning Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EventDefinitionResponse:
      properties:
        event_code:
          type: string
          title: Event Code
        category:
          type: string
          title: Category
        label_en:
          type: string
          title: Label En
        label_ko:
          anyOf:
          - type: string
          - type: 'null'
          title: Label Ko
        label_vi:
          anyOf:
          - type: string
          - type: 'null'
          title: Label Vi
        required_fields:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Required Fields
        regional_applicability:
          type: string
          title: Regional Applicability
      type: object
      required:
      - event_code
      - category
      - label_en
      - label_ko
      - label_vi
      - required_fields
      - regional_applicability
      title: EventDefinitionResponse
...
    FarrowingResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        farm_id:
          type: string
          format: uuid
          title: Farm Id
        sow_id:
          type: string
          format: uuid
          title: Sow Id
        mating_id:
          type: string
          format: uuid
          title: Mating Id
        farrowing_date:
          type: string
          format: date
          title: Farrowing Date
        total_born:
          type: integer
          title: Total Born
        born_alive:
          type: integer
          title: Born Alive
        stillborn:
          type: integer
          title: Stillborn
        mummified:
          type: integer
          title: Mummified
        farrowing_ease:
          anyOf:
          - type: string
          - type: 'null'
          title: Farrowing Ease
        breeding_cycle_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Breeding Cycle Id
        created_at:
          type: string
          format: date-time
          title: Created At
        insights:
          items:
            $ref: '#/components/schemas/EventInsight'
          type: array
          title: Insights
          default: []
      type: object
      required:
      - id
      - farm_id
      - sow_id
      - mating_id
      - farrowing_date
      - total_born
      - born_alive
      - stillborn
      - mummified
      - farrowing_ease
      - breeding_cycle_id
      - created_at
      title: FarrowingResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MatingCreate:
      properties:
        sow_id:
          type: string
          format: uuid
          title: Sow Id
        mating_date:
          type: string
          format: date
          title: Mating Date
        mating_type:
          type: string
          pattern: ^(AI|NATURAL)$
          title: Mating Type
        boar_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Boar Id
        semen_batch:
          anyOf:
          - type: string
          - type: 'null'
          title: Semen Batch
        mating_number:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Mating Number
          default: 1
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - sow_id
      - mating_date
      - mating_type
      title: MatingCreate
    MatingResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        farm_id:
          type: string
          format: uuid
          title: Farm Id
        sow_id:
          type: string
          format: uuid
          title: Sow Id
        mating_date:
          type: string
          format: date
          title: Mating Date
        mating_type:
          type: string
          title: Mating Type
        boar_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Boar Id
        mating_number:
          type: integer
          title: Mating Number
        breeding_cycle_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Breeding Cycle Id
        created_at:
          type: string
          format: date-time
          title: Created At
        insights:
          items:
            $ref: '#/components/schemas/EventInsight'
          type: array
          title: Insights
          default: []
      type: object
      required:
      - id
      - farm_id
      - sow_id
      - mating_date
      - mating_type
      - boar_id
      - mating_number
      - breeding_cycle_id
      - created_at
      title: MatingResponse
...
    WeaningCreate:
      properties:
        sow_id:
          type: string
          format: uuid
          title: Sow Id
        farrowing_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Farrowing Id
        weaning_date:
          type: string
          format: date
          title: Weaning Date
        weaned_count:
          type: integer
          maximum: 30.0
          minimum: 0.0
          title: Weaned Count
        avg_weaning_weight_kg:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Avg Weaning Weight Kg
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - sow_id
      - weaning_date
      - weaned_count
      title: WeaningCreate
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://api.pigos.io
  description: Production
- url: http://localhost:8000
  description: Local dev (iOS 시뮬)
- url: http://10.0.2.2:8000
  description: Local dev (Android 에뮬)